fix rg output when incorrect column is detected
Vims output reader supports either file:linenr:output or
file:linenr:columnnr:output (without column number). The config would
previously request rg to print without a column number, but if one of
the matches starts with 'number:' it will be treated as a column, and
vim will treat all the other output as having a column, which makes them
no longer be resolvable to the correct file location.

This change adds the column number so this can no longer occur, even
though it's slightly longer unnecessarily.